refactor: move hooks tests
