+++ site/src/hooks/use_station_page.test.ts
@@ -1,8 +1,8 @@
-import { useStationPage } from '@hooks/use_station_page'
+import { useStationPage } from './use_station_page'
 
-import { describe, it, expect, vi, beforeAll } from 'vitest'
+import { beforeAll, describe, expect, it, vi } from 'vitest'
 
-import { renderHook, act } from '@testing-library/react'
+import { act, renderHook } from '@testing-library/react'
 
 describe('use station page', () => {
   beforeAll(() => {
